updated airflow setup and dag script to stream data to kafka
