+++ docker-compose.yml
@@ -79,6 +79,63 @@ services:
       PORT: 9021
     networks:
       - confluent
+  
+  airflow-webserver:
+    image: apache/airflow:2.6.0-python3.9
+    command: webserver
+    entrypoint: ['/opt/airflow/script/entrypoint.sh']
+    depends_on: 
+      - postgres
+    environment:
+      LOAD_EX: n
+      EXECUTOR: Sequential
+      AIRFLOW__DATABASE__SQL_ALCHEMY_CONN: postgresql+psycopg2://airflow:airflow@postgres/airflow
+    logging: 
+      options:
+        max-size: 10m
+        max-file: "3"
+    volumes:
+      - ./dags:/opt/airflow/dags
+      - ./scripts/entrypoint.sh:/opt/airflow/script/entrypoint.sh
+    ports:
+      - "8080:8080"
+    healthcheck:
+      test: ['CMD-SHELL',"[ -f /opt/airflow/airflow-webserver.pid ]"]
+      interval: 30s
+      timeout: 10s
+      retries: 5
+    networks:
+      - confluent
+      
+  postgres:
+    image: postgres:14.0
+    environment: 
+      - POSTGRES_USER=airflow
+      - POSTGRES_PASSWORD=airflow
+      - POSTGRES_DB=airflow
+    logging: 
+      options:
+        max-size: 10m
+        max-file: "3"
+    networks:
+      - confluent
+    
+  scheduler:
+    image: apache/airflow:2.6.0-python3.9
+    depends_on: 
+      airflow-webserver:
+        condition: service_healthy 
+    environment:
+      LOAD_EX: n
+      EXECUTOR: Sequential
+      AIRFLOW__DATABASE__SQL_ALCHEMY_CONN: postgresql+psycopg2://airflow:airflow@postgres/airflow
+    volumes:
+      - ./dags:/opt/airflow/dags
+      - ./requirements.txt:/opt/airflow/requirements.txt
+    command: bash -c "pip install -r ./requirements.txt && airflow db upgrade && airflow scheduler"
+    networks:
+      - confluent
+  
 
 networks:
   confluent: {}
